# [#427] [Client] Improve prepare-request-headers performance x3