--- a/src/org/httpkit/client.clj
+++ b/src/org/httpkit/client.clj
@@ -25,11 +25,11 @@
 
 (defn- prepare-request-headers
   [{:keys [headers form-params basic-auth oauth-token user-agent] :as req}]
-  (merge headers
-         (when form-params {"Content-Type"  "application/x-www-form-urlencoded"})
-         (when basic-auth  {"Authorization" (basic-auth-value basic-auth)})
-         (when oauth-token {"Authorization" (str "Bearer " oauth-token)})
-         (when user-agent  {"User-Agent"    user-agent})))
+  (cond-> headers
+    form-params (assoc "Content-Type"  "application/x-www-form-urlencoded")
+    basic-auth  (assoc "Authorization" (basic-auth-value basic-auth))
+    oauth-token (assoc "Authorization" (str "Bearer " oauth-token))
+    user-agent  (assoc "User-Agent"    user-agent)))
 
 (defn- prepare-response-headers [headers]
   (reduce (fn [m [k v]] (assoc m (keyword k) v)) {} headers))